--- src/jigsaw/pc-components/select/select-collapse.html
@@ -1,79 +1,0 @@
-<jigsaw-combo-select
-    width="100%"
-    [autoWidth]="true"
-    [maxHeight]="maxHeight"
-    [(value)]="_$selectedItems"
-    [clearable]="clearable"
-    [labelField]="labelField"
-    [valid]="valid"
-    [disabled]="disabled"
-    [placeholder]="placeholder"
-    [showValueBorder]="multipleSelect"
-    [openTrigger]="openTrigger"
-    [closeTrigger]="closeTrigger"
-    [autoClose]="!multipleSelect"
-    (openChange)="_$onComboOpenChange($event)"
-    (valueChange)="_$handleClearable($event)"
-    (remove)="_$onTagRemove($event)"
->
-    <ng-template>
-        <jigsaw-list
-            [ngClass]="{'jigsaw-select-collapse-list-filter': _$listFilter,
-                        'jigsaw-select-collapse-multiple-list': multipleSelect,
-                        'jigsaw-select-collapse-single-list': !multipleSelect}"
-            [data]="data"
-            [labelField]="labelField"
-            [trackItemBy]="trackItemBy"
-            n
-            [multipleSelect]="multipleSelect"
-            [optionCount]="optionCount"
-            [(selectedItems)]="_$selectedItems"
-            (selectedItemsChange)="_$handleSelectChange($event)"
-        >
-            <div *ngIf="multipleSelect" class="jigsaw-select-collapse-select-all">
-                <jigsaw-checkbox [(checked)]="_$selectAllChecked" *ngIf="!_$listFilter" (checkedChange)="_$selectAll()">
-                    <span>全选</span>
-                </jigsaw-checkbox>
-                <div class="jigsaw-select-collapse-show-selected" *ngIf="!_$listFilter" (click)="_$listShowSelected()">
-                    <span>查看已选</span>
-                </div>
-                <div class="jigsaw-select-collapse-show-all" *ngIf="_$listFilter" (click)="_$listShowAll()">
-                    <i class="iconfont iconfont-e9ac"></i>
-                    <span>返回</span>
-                </div>
-            </div>
-            <jigsaw-collapse class="jigsaw-select-collapse-collapse">
-                <jigsaw-collapse-pane [isActive]="true" *ngFor="let group of data">
-                    <div j-title class="jigsaw-select-collapse-name">{{group[groupName]}}</div>
-                    <div j-body>
-                        <jigsaw-list-option
-                            #listItem
-                            *ngFor="let item of group.data"
-                            [value]="item"
-                            [disabled]="item.disabled"
-                        >
-                            <jigsaw-checkbox
-                                *ngIf="multipleSelect"
-                                [disabled]="item.disabled"
-                                [(checked)]="listItem.selected"
-                            ></jigsaw-checkbox>
-                            <span>{{item && item[labelField] ? item[labelField] : item}}</span>
-                        </jigsaw-list-option>
-                    </div>
-                </jigsaw-collapse-pane>
-            </jigsaw-collapse>
-        </jigsaw-list>
-    </ng-template>
-    <div
-        *ngIf="multipleSelect && overall && _$selectedItems?.length !== 0 && _$selectedItems?.length === _$allOptions?.length"
-        class="jigsaw-select-collapse-overall"
-    >
-        全部
-    </div>
-    <div
-        *ngIf="multipleSelect && overall && _$selectedItems && _$selectedItems?.length !== 0 && _$selectedItems?.length !== _$allOptions?.length"
-        class="jigsaw-select-collapse-overall"
-    >
-        {{"已选择" + _$selectedItems?.length + "项"}}
-    </div>
-</jigsaw-combo-select>
