--- src/jigsaw/pc-components/select/select-collapse.scss
@@ -1,148 +1,0 @@
-@import "../../common/assets/scss/core/base";
-
-$jigsaw-select-collapse: #{$jigsaw-prefix}-select-collapse;
-$jigsaw-combo-select: #{$jigsaw-prefix}-combo-select;
-$jigsaw-collapse: #{$jigsaw-prefix}-collapse;
-$jigsaw-tag: #{$jigsaw-prefix}-tag;
-
-.#{$jigsaw-select-collapse}-host {
-    @include inline-block();
-
-    &.#{$jigsaw-select-collapse}-single {
-        .#{$jigsaw-combo-select}-host {
-            .#{$jigsaw-tag}-host {
-                padding: 0;
-                max-width: 100%;
-                background: none;
-                pointer-events: none;
-            }
-        }
-    }
-
-    &.#{$jigsaw-select-collapse}-multiple {
-        .#{$jigsaw-combo-select}-host {
-            position: relative;
-            .#{$jigsaw-select-collapse}-overall {
-                display: flex;
-                align-items: center;
-                position: absolute;
-                top: 0;
-                width: 100%;
-                height: 100%;
-                padding-left: 8px;
-                pointer-events: none;
-            }
-        }
-        &.#{$jigsaw-select-collapse}-show-overall {
-            .#{$jigsaw-combo-select} {
-                .#{$jigsaw-combo-select}-selection {
-                    .#{$jigsaw-combo-select}-selection-rendered {
-                        jigsaw-tag {
-                            display: none;
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-.#{$jigsaw-select-collapse}-single-list {
-    jigsaw-list-option {
-        padding-left: 16px;
-    }
-}
-
-.#{$jigsaw-select-collapse}-multiple-list {
-    .#{$jigsaw-select-collapse}-select-all {
-        display: flex;
-        align-items: center;
-        height: 32px;
-        border-bottom: 1px solid $splitline-default;
-        cursor: pointer;
-        jigsaw-checkbox {
-            margin: 0 16px 0 8px;
-            .jigsaw-checkbox-wrapper {
-                .jigsaw-checkbox {
-                    margin-right: 6px;
-                }
-                .jigsaw-checkbox-label {
-                    margin: 0;
-                    padding: 0;
-                }
-            }
-        }
-        .#{$jigsaw-select-collapse}-show-selected {
-            display: flex;
-            align-items: center;
-            color: $brand-default;
-            user-select: none;
-            cursor: pointer;
-            &:hover {
-                color: $brand-active;
-            }
-        }
-        .#{$jigsaw-select-collapse}-show-all {
-            display: flex;
-            align-items: center;
-            padding: 0 0 0 8px;
-            color: $brand-default;
-            i {
-                margin-right: 4px;
-                font-size: $icon-size-sm;
-            }
-            &:hover {
-                color: $brand-active;
-            }
-        }
-    }
-
-    jigsaw-list-option {
-        padding-left: 16px;
-        .jigsaw-list-option-wrapper {
-            display: flex;
-            align-items: center;
-            .jigsaw-list-option-title-box {
-                display: none;
-            }
-
-            jigsaw-checkbox {
-                margin-right: 8px;
-            }
-        }
-
-        &.jigsaw-list-option-active {
-            .jigsaw-list-option-wrapper {
-                span {
-                    color: $brand-default;
-                }
-            }
-        }
-    }
-    &.#{$jigsaw-select-collapse}-list-filter {
-        jigsaw-list-option:not(.jigsaw-list-option-active) {
-            display: none;
-        }
-    }
-}
-
-.#{$jigsaw-select-collapse}-collapse {
-    .#{$jigsaw-collapse} {
-        .#{$jigsaw-collapse}-pane {
-            .#{$jigsaw-collapse}-item {
-                .#{$jigsaw-collapse}-header {
-                    padding-left: 8px;
-
-                    .#{$jigsaw-select-collapse}-name {
-                        font-size: $font-title-med;
-                        color: $font-color-heading;
-                    }
-                }
-                .#{$jigsaw-collapse}-content {
-                    .#{$jigsaw-collapse}-padding {
-                        padding: 0px;
-                    }
-                }
-            }
-        }
-    }
-}
