--- src/jigsaw/pc-components/select/select-group.scss
@@ -1,152 +1,0 @@
-@import "../../common/assets/scss/core/base";
-
-$jigsaw-select-group: #{$jigsaw-prefix}-select-group;
-$jigsaw-combo-select: #{$jigsaw-prefix}-combo-select;
-$jigsaw-collapse: #{$jigsaw-prefix}-collapse;
-$jigsaw-tag: #{$jigsaw-prefix}-tag;
-
-.#{$jigsaw-select-group}-host {
-    @include inline-block();
-
-    &.#{$jigsaw-select-group}-single {
-        .#{$jigsaw-combo-select}-host {
-            .#{$jigsaw-tag}-host {
-                padding: 0;
-                max-width: 100%;
-                background: none;
-                pointer-events: none;
-            }
-        }
-    }
-
-    &.#{$jigsaw-select-group}-multiple {
-        .#{$jigsaw-combo-select}-host {
-            position: relative;
-            .#{$jigsaw-select-group}-overall {
-                display: flex;
-                align-items: center;
-                position: absolute;
-                top: 0;
-                width: 100%;
-                height: 100%;
-                padding-left: 8px;
-                pointer-events: none;
-            }
-        }
-        &.#{$jigsaw-select-group}-show-overall {
-            .#{$jigsaw-combo-select} {
-                .#{$jigsaw-combo-select}-selection {
-                    .#{$jigsaw-combo-select}-selection-rendered {
-                        jigsaw-tag {
-                            display: none;
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-.#{$jigsaw-select-group}-single-list {
-    jigsaw-list-option {
-        padding-left: 16px;
-    }
-}
-
-.#{$jigsaw-select-group}-multiple-list {
-    .#{$jigsaw-select-group}-select-all {
-        display: flex;
-        align-items: center;
-        height: 32px;
-        border-bottom: 1px solid $splitline-default;
-        cursor: pointer;
-        jigsaw-checkbox {
-            margin: 0 16px 0 8px;
-            .jigsaw-checkbox-wrapper {
-                .jigsaw-checkbox {
-                    margin-right: 6px;
-                }
-                .jigsaw-checkbox-label {
-                    margin: 0;
-                    padding: 0;
-                }
-            }
-        }
-        .#{$jigsaw-select-group}-show-selected {
-            display: flex;
-            align-items: center;
-            color: $brand-default;
-            user-select: none;
-            cursor: pointer;
-            &:hover {
-                color: $brand-active;
-            }
-        }
-        .#{$jigsaw-select-group}-show-all {
-            display: flex;
-            align-items: center;
-            padding: 0 0 0 8px;
-            color: $brand-default;
-            i {
-                margin-right: 4px;
-                font-size: $icon-size-sm;
-            }
-            &:hover {
-                color: $brand-active;
-            }
-        }
-    }
-    jigsaw-list-option {
-        padding-left: 16px;
-        .jigsaw-list-option-wrapper {
-            display: flex;
-            align-items: center;
-            .jigsaw-list-option-title-box {
-                display: none;
-            }
-
-            jigsaw-checkbox {
-                margin-right: 8px;
-            }
-        }
-
-        &.jigsaw-list-option-active {
-            .jigsaw-list-option-wrapper {
-                span {
-                    color: $brand-default;
-                }
-            }
-        }
-    }
-    &.#{$jigsaw-select-group}-list-filter {
-        jigsaw-list-option:not(.jigsaw-list-option-active) {
-            display: none;
-        }
-    }
-}
-
-.#{$jigsaw-select-group}-collapse {
-    .#{$jigsaw-collapse} {
-        .#{$jigsaw-collapse}-pane {
-            .#{$jigsaw-collapse}-item {
-                .#{$jigsaw-collapse}-header {
-                    padding-left: 8px;
-                    pointer-events: none;
-
-                    .#{$jigsaw-collapse}-arrow {
-                        display: none;
-                    }
-
-                    .#{$jigsaw-select-group}-name {
-                        font-size: $font-title-sm;
-                        color: $font-color-watermark;
-                    }
-                }
-                .#{$jigsaw-collapse}-content {
-                    .#{$jigsaw-collapse}-padding {
-                        padding: 0px;
-                    }
-                }
-            }
-        }
-    }
-}
