+++ src/jigsaw/pc-components/theming/all-theme.scss
@@ -24,8 +24,6 @@
 @import "../pagination/pagination-item";
 @import "../radio/radio";
 @import "../select/select";
-@import "../select/select-collapse.scss";
-@import "../select/select-group.scss";
 @import "../slider/slider";
 @import "../steps/step";
 @import "../switch/switch";
